Remove `ScrapePartnersCommandV2`, update related documentation and console commands, and refactor partner scraping for improved base domain handling and progress tracking.
